requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
schedule==1.2.1

# Google API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# API Clients
requests==2.31.0
python-linkedin-v2==0.9.4
google-api-python-client==2.108.0
eventbrite==0.5.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
logging==0.4.9.6 